slack_overflow/requirements.txt
slack-bolt==1.16.4
slack-sdk==3.20.2
python-dotenv==1.0.0
Django==4.1.7
djangorestframework==3.14.0
